+++ src/Controller/ControllerManager.php
@@ -16,6 +16,7 @@ use Zend\Mvc\Exception;
 use Zend\ServiceManager\AbstractPluginManager;
 use Zend\ServiceManager\ConfigInterface;
 use Zend\ServiceManager\Exception\InvalidServiceException;
+use Zend\ServiceManager\ServiceLocatorAwareInterface;
 use Zend\Stdlib\DispatchableInterface;
 
 /**
@@ -54,6 +55,10 @@ class ControllerManager extends AbstractPluginManager
         $this->addInitializer([$this, 'injectConsole']);
         $this->addInitializer([$this, 'injectPluginManager']);
         parent::__construct($configOrContainerInstance, $v3config);
+
+        // Added after parent construction, as v2 abstract plugin managers add
+        // one during construction.
+        $this->addInitializer([$this, 'injectServiceLocator']);
     }
 
     /**
@@ -190,5 +195,52 @@ class ControllerManager extends AbstractPluginManager
         }
 
         $controller->setPluginManager($container->get('ControllerPluginManager'));
+    }
+
+    /**
+     * Initializer: inject service locator
+     *
+     * @param ContainerInterface|DispatchableInterface $first Container when
+     *     using zend-servicemanager v3; controller under v2.
+     * @param DispatchableInterface|ContainerInterface $second Controller when
+     *     using zend-servicemanager v3; container under v2.
+     */
+    public function injectServiceLocator($first, $second)
+    {
+        printf("In %s\n", __METHOD__);
+        if ($first instanceof ContainerInterface) {
+            $container = $first;
+            $controller = $second;
+        } else {
+            $container = $second;
+            $controller = $first;
+        }
+
+        // For v2, we need to pull the parent service locator
+        if (! method_exists($container, 'configure')) {
+            $container = $container->getServiceLocator() ?: $container;
+        }
+
+        if (! $controller instanceof ServiceLocatorAwareInterface
+            && method_exists($controller, 'setServiceLocator')
+        ) {
+            trigger_error(sprintf(
+                'ServiceLocatorAwareInterface is deprecated and will be removed in version 3.0, along '
+                . 'with the ServiceLocatorAwareInitializer. Please update your class %s to remove '
+                . 'the implementation, and start injecting your dependencies via factory instead.',
+                get_class($controller)
+            ), E_USER_DEPRECATED);
+            $controller->setServiceLocator($container);
+        }
+
+        if ($controller instanceof ServiceLocatorAwareInterface) {
+            trigger_error(sprintf(
+                'ServiceLocatorAwareInterface is deprecated and will be removed in version 3.0, along '
+                . 'with the ServiceLocatorAwareInitializer. Please update your class %s to remove '
+                . 'the implementation, and start injecting your dependencies via factory instead.',
+                get_class($controller)
+            ), E_USER_DEPRECATED);
+            $controller->setServiceLocator($container);
+        }
     }
 }